♻️ : 타입스크립트 추상 클래스 적용
