+++ src/es/Store/Interface.ts
@@ -1,9 +1,3 @@
-export interface IStore {
-  addSubscriber(subscriber: object): void;
-  setState(newState: object): void;
-  getState(): object;
-}
-
 export interface IProduct {
   name: string;
   price: number;
