+++ src/es/Store/ProductStore.ts
@@ -1,30 +1,14 @@
-import { IStore, IProduct } from './Interface';
+import { IProduct } from './Interface';
+import Store from './Abstract';
 
 interface IState {
   products: Array<IProduct>;
 }
 
-class ProductStore implements IStore {
-  private state: IState = {
+class ProductStore extends Store {
+  protected state: IState = {
     products: [],
   };
-
-  private subscribers = [];
-
-  addSubscriber(subscriber: object) {
-    this.subscribers.push(subscriber);
-  }
-
-  setState(newState: IState) {
-    const changeStates: Array<string> = Object.entries(newState).map(([key]) => key);
-
-    this.state = { ...this.state, ...newState };
-    this.subscribers.forEach(renderMethod => renderMethod({ state: this.state, changeStates }));
-  }
-
-  getState(): IState {
-    return { ...this.state };
-  }
 
   addProduct(product: IProduct): void {
     this.setState({
